Create feature_generation.py
Currently, only generates pool of features up to complexity 2 (in terms of the recursive def. of complexity).

The general pool is generated by using all object types in the TYPES list. However, each object types T that does not occur in a given domain is equivalent to falsum for that domain. Thus, any predicate using T is equivalent to the predicate using falsum in place of T. Instead of generating the entire pool and pruning all such T-concepts later, we can generate a separate predicate pool for each domain which already avoids all such T-concepts. This is done by  commenting out predicates from TYPES appropriately.